Homework #7
Networking

API KEY добавляется в запросы через ApiKeyInterceptor.
Query параметры методов удалены
