+++ .gitignore
@@ -84,3 +84,4 @@ lint/generated/
 lint/outputs/
 lint/tmp/
 # lint/reports/
+/app/src/main/assets/api_key
