--- app/src/main/assets/people.json
@@ -1,135 +1,0 @@
-[
-  {
-    "gender": 2,
-    "id": 84433,
-    "name": "Lannick Gautry",
-    "original_name": "Lannick Gautry",
-    "profile_path": "https://image.tmdb.org/t/p/w342/h94QUkKrwUncYhJ1eMz23kBAJuc.jpg"
-  },
-  {
-    "gender": 2,
-    "id": 37919,
-    "name": "Stanislas Merhar",
-    "original_name": "Stanislas Merhar",
-    "profile_path": "https://image.tmdb.org/t/p/w342/k5XQM1XXG71GLd41hFcx8yhkxRy.jpg"
-  },
-  {
-    "gender": 2,
-    "id": 1407184,
-    "name": "Kaaris",
-    "original_name": "Kaaris",
-    "profile_path": "https://image.tmdb.org/t/p/w342/zHxqojtMgk5aBqMQyTqcoPc4TRG.jpg"
-  },
-  {
-    "gender": 2,
-    "id": 1003,
-    "name": "Jean Reno",
-    "original_name": "Jean Reno",
-    "profile_path": "https://image.tmdb.org/t/p/w342/q7dYamebioHRuvb9EWeSw8yTEfS.jpg"
-  },
-  {
-    "gender": 2,
-    "id": 62439,
-    "name": "David Belle",
-    "original_name": "David Belle",
-    "profile_path": "https://image.tmdb.org/t/p/w342/nxRnf3O6Y3Yldd0pibuo43x3RZ.jpg"
-  },
-  {
-    "gender": 2,
-    "id": 70165,
-    "name": "Gérard Lanvin",
-    "original_name": "Gérard Lanvin",
-    "profile_path": "https://image.tmdb.org/t/p/w342/3FSr4VW7JF3mBcVwPnJmOeR09NB.jpg"
-  },
-  {
-    "gender": 2,
-    "id": 62,
-    "name": "Bruce Willis",
-    "original_name": "Bruce Willis",
-    "profile_path": "https://image.tmdb.org/t/p/w342/A1XBu3CffBpSK8HEIJM8q7Mn4lz.jpg"
-  },
-  {
-    "gender": 2,
-    "id": 57171,
-    "name": "Jesse Metcalfe",
-    "original_name": "Jesse Metcalfe",
-    "profile_path": "https://image.tmdb.org/t/p/w342/qZR4VAB2t0qawkH88Wv3K7y67nM.jpg"
-  },
-  {
-    "gender": 1,
-    "id": 1734214,
-    "name": "Natalie Eva Marie",
-    "original_name": "Natalie Eva Marie",
-    "profile_path": "https://image.tmdb.org/t/p/w342/nS7tdSPGtmSlkMCHO2W18SgFVWP.jpg"
-  },
-  {
-    "gender": 1,
-    "id": 1798408,
-    "name": "Lala Kent",
-    "original_name": "Lala Kent",
-    "profile_path": "https://image.tmdb.org/t/p/w342/rWhoOelbtMg5iYyDgOAa2IagHMz.jpg"
-  },
-  {
-    "gender": 2,
-    "id": 53185,
-    "name": "Texas Battle",
-    "original_name": "Texas Battle",
-    "profile_path": "https://image.tmdb.org/t/p/w342/zPzagAtKxbHKrQLAQ69KSYLZxjA.jpg"
-  },
-  {
-    "gender": 2,
-    "id": 1075049,
-    "name": "Swen Temmel",
-    "original_name": "Swen Temmel",
-    "profile_path": "https://image.tmdb.org/t/p/w342/z6zLmf3PBsTGMU3RBnguAqfHUn3.jpg"
-  },
-  {
-    "gender": 2,
-    "id": 78798,
-    "name": "Tom Kenny",
-    "original_name": "Tom Kenny",
-    "profile_path": "https://image.tmdb.org/t/p/w342/o6sC2IsjL7KncT4pZBOx4T0WKhM.jpg"
-  },
-  {
-    "gender": 2,
-    "id": 34398,
-    "name": "Bill Fagerbakke",
-    "original_name": "Bill Fagerbakke",
-    "profile_path": "https://image.tmdb.org/t/p/w342/sx3LkX82tpdQfGi3ie3cFvrgCYg.jpg"
-  },
-  {
-    "gender": 2,
-    "id": 70615,
-    "name": "Rodger Bumpass",
-    "original_name": "Rodger Bumpass",
-    "profile_path": "https://image.tmdb.org/t/p/w342/9TEh5QJAHnXuM3fE1Uu40QzXJQC.jpg"
-  },
-  {
-    "gender": 2,
-    "id": 188229,
-    "name": "Mr. Lawrence",
-    "original_name": "Mr. Lawrence",
-    "profile_path": "https://image.tmdb.org/t/p/w342/eLI17X32n91iMuLroXtpKUlMOQf.jpg"
-  },
-  {
-    "gender": 2,
-    "id": 6384,
-    "name": "Keanu Reeves",
-    "original_name": "Keanu Reeves",
-    "profile_path": "https://image.tmdb.org/t/p/w342/rRdru6REr9i3WIHv2mntpcgxnoY.jpg"
-  },
-  {
-    "gender": 2,
-    "id": 6574,
-    "name": "Clancy Brown",
-    "original_name": "Clancy Brown",
-    "profile_path": "https://image.tmdb.org/t/p/w342/xjg0ZIxP0tFcEQCTeRgKxNtLdpe.jpg"
-  },
-  {
-    "gender": 1,
-    "id": 78799,
-    "name": "Carolyn Lawrence",
-    "original_name": "Carolyn Lawrence",
-    "profile_path": "https://image.tmdb.org/t/p/w342/7Ehx4bEO8loovbmbntXl0sBNjtj.jpg"
-  }
-]
