+++ app/src/main/java/com/example/superpupermegaproject/App.kt
@@ -1,26 +1,43 @@
 package com.example.superpupermegaproject
 
 import android.app.Application
-import com.example.superpupermegaproject.model.RESTApiInteractor
 import com.example.superpupermegaproject.model.Repository
+import com.example.superpupermegaproject.model.MoviesInteractor
 
-class App: Application() {
-    companion object {
-        private var repositoryInstance: Repository? = null
-        private var APIInteractorInstance: RESTApiInteractor? = null
+class App : Application() {
+    override fun onCreate() {
+        super.onCreate()
 
-        fun getRepositoryInstance() : Repository = if(repositoryInstance==null) {
-            Repository.getInstance(getAPIInteractorInstance())
-        } else {
-            repositoryInstance!!
+        initApiKey()
+    }
+
+    private fun initApiKey() {
+        try {
+            val stream = assets.open("api_key")
+            apiKey = stream.reader().readText().trim()
+            stream.close()
+        } catch (t: Throwable) {
+            throw Exception("Не доступен файл api_key в каталоге assets. ${t.localizedMessage}")
         }
+    }
 
+    companion object {
+        private var apiKey: String = ""
+        private var moviesInteractorInstance: MoviesInteractor? = null
+        private var repositoryInstance: Repository? = null
 
-        fun getAPIInteractorInstance(): RESTApiInteractor =
-            if(APIInteractorInstance==null) {
-                RESTApiInteractor()
+        fun getMoviesInteractorInstance(): MoviesInteractor =
+            if (moviesInteractorInstance == null) {
+                MoviesInteractor.getInstance(getRepositoryInstance())
             } else {
-                APIInteractorInstance!!
+                moviesInteractorInstance!!
+            }
+
+        private fun getRepositoryInstance(): Repository =
+            if (repositoryInstance == null) {
+                Repository(apiKey)
+            } else {
+                repositoryInstance!!
             }
     }
 }
