+++ app/src/main/java/com/example/superpupermegaproject/ui/viewmodels/MovieDetailsViewModel.kt
@@ -1,16 +1,16 @@
-package com.example.superpupermegaproject.ui
+package com.example.superpupermegaproject.ui.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.example.superpupermegaproject.data.Movie
-import com.example.superpupermegaproject.model.Repository
+import com.example.superpupermegaproject.model.MoviesInteractor
 
-class MovieDetailsViewModel(val repository: Repository) : ViewModel() {
+class MovieDetailsViewModel(val moviesInteractor: MoviesInteractor) : ViewModel() {
     private val _movieObservable = MutableLiveData<Movie>()
     val movieObservable: LiveData<Movie> = _movieObservable
 
     suspend fun getMovie(id: Long) {
-        _movieObservable.value = repository.getMovie(id.toInt())
+        _movieObservable.value = moviesInteractor.getMovie(id.toInt())
     }
 }
