+++ app/src/main/java/com/example/superpupermegaproject/ui/viewmodels/ViewModelFactory.kt
@@ -1,0 +1,14 @@
+package com.example.superpupermegaproject.ui.viewmodels
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import com.example.superpupermegaproject.model.MoviesInteractor
+
+class ViewModelFactory(val moviesInteractor: MoviesInteractor): ViewModelProvider.NewInstanceFactory() {
+    override fun <T : ViewModel?> create(modelClass: Class<T>): T =
+        when(modelClass) {
+            MoviesListViewModel::class.java -> MoviesListViewModel(moviesInteractor) as T
+            MovieDetailsViewModel::class.java -> MovieDetailsViewModel(moviesInteractor) as T
+            else -> throw(Exception("Некорректный класс ViewModel"))
+        }
+}
